Se completo vista de facturas y ventana modal para el detalle

+++ Backend/licorstore/src/main/java/com/aglayatech/licorstore/repository/IFacturaRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.aglayatech.licorstore.model.Factura;
 
-public interface IFacturaRepository extends JpaRepository<Factura, Integer> {
+public interface IFacturaRepository extends JpaRepository<Factura, Long> {
 
 }
